Move and expand Parameters and Attribute tests
Renamed test files to the DataModel directory for better organization. Expanded ParametersTest with additional test cases and removed the old ParametersTest.php.
